requirements.txt
aiofiles==23.1.0
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
black==23.3.0
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
frozenlist==1.3.3
idna==3.4
loguru==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
orjson==3.8.9
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
py-cord==2.4.1
pycparser==2.21
pyi18n-v2==1.2.1
PyNaCl==1.5.0
python-decouple==3.7
PyYAML==6.0
tomli==2.0.1
typing_extensions==4.4.0
yarl==1.8.2
